+++ app/src/main/java/wildbakery/ufu/Fragments/DetailFragmentNews.java
