+++ app/src/main/java/wildbakery/ufu/Fragments/FragmentNews.java
