+++ app/src/main/java/wildbakery/ufu/Models/NewsItem.java
