+++ app/src/main/java/wildbakery/ufu/Models/QueryModel.java
